Moves file tools to data_upload module
