+++ broker/data_upload/schema/ztf_3.2.json
